[SmartPlaylist] Move RFC3339 full-date date validation into CDateTime
